requirements.txt
python-dotenv==0.19.2
windows-curses==2.3.0
